+++ apps/connect/src/components/app.tsx
@@ -1,4 +1,9 @@
-import { Analytics, Router } from "@powerhousedao/connect/components";
+import {
+  Analytics,
+  OpenPanel,
+  Router,
+} from "@powerhousedao/connect/components";
+
 import { SentryProvider } from "@powerhousedao/connect/context";
 import {
   DocumentEditorDebugTools,
@@ -58,6 +63,7 @@ export const App = () => {
       <Router />
       <PackageInstallPrompt />
       <Analytics />
+      <OpenPanel />
     </SentryProvider>
   );
 };